refactor: OutputView 내에서 출력 담당 변경
